pyproject.toml
[project]
name = "weewx_axis"
description = "WeeWX extension to overlay weather data on an Axis camera stream"
version = "0.0.1"
authors = [{name="Liz Stanley", email="lizmstanley@gmail.com"}]
maintainers = [{name="Liz Stanley"}]
readme = "README.md"
requires-python = ">=3.14,<4.0"
dependencies = [
    "python-dotenv>=1.2.1",
    "toml>=0.10.2",
    "weewx>=5.2.0",
    "requests>=2.32.5"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"